requirements.txt
streamlit==1.30.0
streamlit-webrtc==0.35.0
opencv-python-headless==4.5.4.60
pyzbar==0.1.8
Pillow==8.4.0
numpy==1.21.4
requests==2.26.0